% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{all_permutations}
\alias{all_permutations}
\title{Get all size K permutations of `0` and `1`}
\usage{
all_permutations(K)
}
\arguments{
\item{K}{number of dimensions}
}
\value{
a data frame of permutations
}
\description{
Get all size K permutations of `0` and `1`
}
\examples{
all_permutations(3)
}
