% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{bivariateCondProb}
\alias{bivariateCondProb}
\title{Extract Theoretical Conditional Probabilities when all attention is set on kstar}
\usage{
bivariateCondProb(parameters, kstar)
}
\arguments{
\item{parameters}{an objection as created by `get_params()`}

\item{kstar}{the dimension k with all attention}
}
\value{
a vector of conditional probabilities
}
\description{
Extract Theoretical Conditional Probabilities when all attention is set on kstar
}
