% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{calculateDistSim}
\alias{calculateDistSim}
\title{Calculate distances and similarity from features to prototype}
\usage{
calculateDistSim(P, w, data, g, r = 1)
}
\arguments{
\item{P}{a K-sized vector of binary features (i.e., the prototype)}

\item{w}{a K-sized vector of attention weights}

\item{data}{a data frame, as returned by the `make_data()` function}

\item{g}{a sensitivity parameter, a number larger than zero}

\item{r}{the type of distance, 1 for Manhattan, 2 for Euclidean}
}
\value{
a data frame with distance and similarity
}
\description{
Calculate distances and similarity from features to prototype
}
