% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateDS.R
\name{calculateDS}
\alias{calculateDS}
\title{Calculate distances and similarity from features to prototype}
\usage{
calculateDS(X, w, P, g, r = 1)
}
\arguments{
\item{X}{a matrix, as returned by the `make_data()` function}

\item{w}{a K-sized vector of attention weights}

\item{P}{a K-sized vector of binary features (i.e., the prototype)}

\item{g}{a sensitivity parameter, a number larger than zero}

\item{r}{the type of distance, 1 for Manhattan, 2 for Euclidean}
}
\value{
a data frame with distance and similarity
}
\description{
Calculate distances and similarity from features to prototype
}
