% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositionalEffects.R
\name{compositionalEffect}
\alias{compositionalEffect}
\title{Compositional Effects}
\usage{
compositionalEffect(reference, comparison)
}
\arguments{
\item{reference}{a simulation, as provided by the `compute()` function}

\item{comparison}{another simulation, as provided by the `compute()` function}
}
\value{
a list of data frames showing difference in probabilities and relative risk ratios
}
\description{
Compositional Effects
}
