% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{compute}
\alias{compute}
\title{Calculate distance, similarity, and membership probabilities}
\usage{
compute(prototypes, w, data, g, r = 1L)
}
\arguments{
\item{prototypes}{a list of prototypes, i.e., K-sized vectors of binary
features.}

\item{w}{a K-sized vector of attention weights}

\item{data}{a data frame, as returned by the `make_data()` function}

\item{g}{a sensitivity parameter, a number larger than zero}

\item{r}{the type of distance, 1 for Manhattan, 2 for Euclidean}
}
\value{
A `prototype` object. A list of probabilities, similarities,
 distances, and the data used to calculate them.
}
\description{
Calculate distance, similarity, and membership probabilities
}
