% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{conditionalProbs}
\alias{conditionalProbs}
\title{Get Conditional Probabilities}
\usage{
conditionalProbs(x, type = c("features", "categories"), s = 500)
}
\arguments{
\item{x}{a `prototype` object created by the `compute()` function.}

\item{type}{whether "features" `Pr(X|C)` or "categories" `Pr(C|X)`}

\item{s}{number of draws to sample from the `.$probabilities` object created
by the `compute()` function}
}
\value{
a list of conditional probabilities
}
\description{
Get Conditional Probabilities
}
