% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{conditionalProbsWhichMax}
\alias{conditionalProbsWhichMax}
\title{Get Conditional Probabilities for K Features}
\usage{
conditionalProbsWhichMax(x)
}
\arguments{
\item{x}{a `prototype` object created by the `compute()` function.}
}
\value{
a list of conditional probabilities; e.g., `Pr(K = 1 | C = 1)`
}
\description{
Get Conditional Probabilities for K Features
}
