% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_data_params}
\alias{get_data_params}
\title{Extract parameters from prototypeData object}
\usage{
get_data_params(x)
}
\arguments{
\item{x}{A prototypeData object created by \code{\link{make_binary_data}}}
}
\value{
A list with two components:
  \describe{
    \item{marginals}{Named numeric vector of marginal probabilities}
    \item{rho}{Correlation matrix used to generate the data}
  }
}
\description{
Retrieves the parameters (marginal probabilities and correlation matrix)
that were used to generate a prototypeData object.
}
\seealso{
\code{\link{make_binary_data}}
}
