% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_params}
\alias{get_params}
\title{Extract parameters from `prototypeData` object}
\usage{
get_params(x)
}
\arguments{
\item{x}{a `prototypeData` object}
}
\value{
a list with (1) the marginal probabilities of x and (2) the correlation matrix of x
}
\description{
Extract parameters from `prototypeData` object
}
