% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{rlkjcorr}
\alias{rlkjcorr}
\title{LKJ distribution}
\source{
These functions come from Richard McElreath's `rethinking` package
and, judging by the source code, it seems that he got them from Ben Goodrich.
}
\usage{
rlkjcorr(n, K, eta = 1)
}
\arguments{
\item{n}{number of draws}

\item{K}{number of rows and columns in the matrix}

\item{eta}{prior, as eta increases more mass is placed over identity matrices}
}
\value{
a matrix or list of matrices
}
\description{
LKJ distribution
}
