% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{temperature}
\alias{temperature}
\title{Temperature Scaling}
\usage{
temperature(w, temp = 1)
}
\arguments{
\item{w}{numeric vector of probabilities (must sum to 1)}

\item{temp}{numeric, temperature parameter
- temp > 1: makes distribution more uniform
- temp < 1: makes distribution less uniform
- temp = 1: no change}
}
\value{
numeric vector of temperature-scaled probabilities
}
\description{
Temperature Scaling
}
\examples{

w <- runif(10)
w <- w / sum(w)

temperature(w, temp = 0)
temperature(w, temp = 0.5)
temperature(w, temp = 1)

}
