% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{conditionalProbs}
\alias{conditionalProbs}
\title{Get Conditional Probabilities for K Features}
\usage{
conditionalProbs(x, .sample = FALSE)
}
\arguments{
\item{x}{a `prototype` object created by the `compute()` function.}

\item{.sample}{an indicator that determines whether the classification is done
using `which.max` (`.sample = FALSE`) or using `sample` (`.sample = TRUE`)}
}
\value{
a list of conditional probabilities; e.g., `Pr(K = 1 | C = 1)`
}
\description{
Get Conditional Probabilities for K Features
}
