% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{conditionalProbsSample}
\alias{conditionalProbsSample}
\title{Get Conditional Probabilities for K Features}
\usage{
conditionalProbsSample(x, s = 500)
}
\arguments{
\item{x}{a `prototype` object created by the `compute()` function.}

\item{s}{number of draws to sample from the `.$probabilities` object created
by the `compute()` function}
}
\value{
a list of conditional probabilities; e.g., `Pr(K = 1 | C = 1)`
}
\description{
Get Conditional Probabilities for K Features
}
