% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate computation into a single data frame}
\usage{
consolidate(x)
}
\arguments{
\item{x}{a `prototype` object created by the `compute()` function.}
}
\value{
a data frame
}
\description{
Consolidate computation into a single data frame
}
