% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_binary_data}
\alias{make_binary_data}
\title{Make Correlated Binary Data}
\usage{
make_binary_data(marginals, rho, obs = 1000)
}
\arguments{
\item{marginals}{A numeric vector of marginal probabilities for each variable.}

\item{rho}{A symmetric correlation matrix with dimensions matching the length
of \code{marginals}}

\item{obs}{Integer. Number of observations (rows) to generate.}
}
\value{
A \code{prototypeData} object containing:
  \describe{
    \item{Binary data}{A data frame with \code{obs} rows and \code{length(marginals)} columns}
    \item{params attribute}{List containing the original marginals and correlation matrix}
  }
}
\description{
Generates correlated binary data. The function first generates multivariate
normal data with specified correlations, then transforms it to binary data while preserving the
correlation structure. Apparently this is known as a "Gaussian copula" approach.
}
\examples{
# Generate 8-dimensional correlated binary data
K <- 8
marginals <- rbeta(K, 2, 3)
rho <- rlkjcorr(1, K, eta = 1 / 4)
out <- make_binary_data(marginals, rho)
out

}
\seealso{
\code{\link{get_data_params}}, \code{\link{print.prototypeData}}
}
