% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{make_binary_data}
\alias{make_binary_data}
\title{Make Binary Data}
\usage{
make_binary_data(marginals, rho, obs = 1000)
}
\arguments{
\item{marginals}{a K-sized vector of marginal probabilities}

\item{rho}{a correlation matrix}

\item{obs}{number of rows in the dataset}
}
\value{
a matrix of simulated observations
}
\description{
Make Binary Data
}
\examples{
K <- 8
marginals <- rbeta(K, 2, 3)
rho <- rlkjcorr(1, K, eta = 1 / 4)
out <- make_binary_data(marginals, rho)
head(out)
colMeans(out)
marginals
}
