% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{transform_rho}
\alias{transform_rho}
\title{Transform correlation}
\usage{
transform_rho(rho, el)
}
\arguments{
\item{rho}{correlation matrix}

\item{el}{a K-by-3 matrix in which the first two columns are row
and column indicators; and the third column is the new value.}
}
\value{
a transformed correlation matrix, further adjusted so that
   it is "positive definite"
}
\description{
Transform correlation
}
\examples{

rho <- rlkjcorr(n = 1, K = 4, eta = 1)

el <- rbind(
  c(i = 1, j = 2, r = +0.7), ## set corr between x1 and x2 as 0.7
  c(i = 3, j = 4, r = -0.5)  ## set corr between x3 and x4 as -0.5
)

out <- transform_rho(rho, el)

round(rho, 2)
round(out, 2)

}
